fix import, replace TestT name, split test cases
